Theme initialization task
In the existing document, the command to copy the configuration file with command cp, which does not consider the Windows user.
Since there is no way to determine the OS in the npm script, I made it into a separate js file and run it on the node.

There is also a problem that when the user pushes files to their github page,
the theme does not come up in github,
due to the .git directory existing in this theme directory.
Therefore, init.js also deletes the .git directory.

I had to delete the .gitignore in init.js because the .gitignore file did not get the details of the theme changes.