Add font
Previously, we had to install noto-sans-kr, spoqa-han-sans-kr, and spoqa-han-sans-jp fonts. I also clone all font related files in the spoqa-han-sans repository, so I
need to get hundreds of megabytes of files, so I only get the files related to web fonts.

Also, when cloning the repository, there is a problem that the .git directory is created and does not go to the github repository, so we have to delete the .git direc
tories inside the font directory in rm-git.js.