PR for v1.0.0 (#45)
* change .gitignore for IntelliJ

* added that README.md only supports node 7 or below.

hexo-renderer-bourbon v1.0.4 isn't support Node v8.

* Theme initialization task

In the existing document, the command to copy the configuration file with command cp, which does not consider the Windows user.
Since there is no way to determine the OS in the npm script, I made it into a separate js file and run it on the node.

There is also a problem that when the user pushes files to their github page,
the theme does not come up in github,
due to the .git directory existing in this theme directory.
Therefore, init.js also deletes the .git directory.

I had to delete the .gitignore in init.js because the .gitignore file did not get the details of the theme changes.

* Add font

Previously, we had to install noto-sans-kr, spoqa-han-sans-kr, and spoqa-han-sans-jp fonts. I also clone all font related files in the spoqa-han-sans repository, so I
need to get hundreds of megabytes of files, so I only get the files related to web fonts.

Also, when cloning the repository, there is a problem that the .git directory is created and does not go to the github repository, so we have to delete the .git direc
tories inside the font directory in rm-git.js.

* Update package.json

* fix error of rel attribute value

alternative -> alternate

* fix error for npm script

split build task by each command and re-integrate

* fix accessibility

remove limitted maximum-scale from viewport setting

* fix the difference between canonical url and opengraph url

* apply 100% height layout

* add LiveRe Platform (소셜덧글 플랫폼)

* Migrate SASS to PostCSS 👼

And, FINALLY remove the bourbon.
This commit fixes #35, and #44.

* Embed font and icon files

Now users don't need to clone repos. So I delete commands which clone repos.
This fixes #30.

* Delete unneeded scripts

* Change .jade to .pug

* Update README

* Fix some bugs

* Fix page indicator margin

* Fix syntax which is not supported in pug anymore

* Fix missing `&` operators

* Add githook for build process

*  Add missing `&` operator

* Change built CSS file

* Fix some bugs, Improve article layout

* Add extra data for twitter and slack

* Improve meta:description

* Fix <hr> style bug